feat: move some token processing to the lexer
